+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 🪄 Suggest missing import whenever possible
 - ️🧐 Marking self import as invalid
 - ️🧐 Marking duplicated type name as invalid
+- ️🧐 Marking empty and invalid import as invalid
 
 ### Fixed
 - Added missing top level keyword import